Add some sort of auth

+++ src/service/delete-message.go
@@ -9,6 +9,11 @@ import (
 
 func (s *service) DeleteMessage(context echo.Context) error {
 	clientId := context.Param("client_id")
+	clientIdAuth := context.Request().Header.Get("client_id")
+
+	if clientId != clientIdAuth {
+		return echo.NewHTTPError(http.StatusUnauthorized, "Client not authorized to delete message")
+	}
 	
 	err := s.repository.Delete(clientId)
 	if err != nil {
